+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,0 +1,1 @@
+export { keyringController } from './keyring-controller';